+++ backend/problems/src/ma1/equations/inequalities.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use macros::problem;
 use math::{
-    MathDisplay,
+    MathDisplay, formatting,
     num_gen::{self, NumberGenerator},
     symbols::{self, inequality_sign::InequalitySign},
 };
@@ -11,7 +11,6 @@ use types::{
     lang::Language,
     problems::{Problem, ProblemParameters, Solution},
 };
-use typst_writer::formatting;
 
 /// x - 5 > 8
 /// Absolute difficulty: 3